chore: update sendinblue api

+++ investment/management/commands/check_inactive_users.py
@@ -1,11 +1,13 @@
 from datetime import datetime, timedelta
 
+from django.contrib.auth import get_user_model
+
 from django.conf import settings
 from django.core.management import BaseCommand
 
 from investment.models import Investment
 
-User = settings.AUTH_USER_MODEL
+User = get_user_model()
 
 
 class Command(BaseCommand):
@@ -14,7 +16,10 @@ class Command(BaseCommand):
     def handle(self, *args, **kwargs):
         users = User.objects.filter(is_active=False)
         today = datetime.today
-        for user in users:
-            if user.start_date + timedelta(days=3) > today:
-                user.delete()
-                self.stdout.write(f"User {user.username} deleted successfully!")
+        if users:
+            for user in users:
+                if user.start_date + timedelta(days=3) > today:
+                    user.delete()
+                    self.stdout.write(f"User {user.username} deleted successfully!")
+        else:
+            self.stdout.write("No Inactive Users Available")
